requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
gunicorn
google-genai
Pillow>=10.0.0
